+++ src/Microsoft.DotNet.ImageBuilder/src/Commands/CopyDockerHubBaseImagesOptions.cs
@@ -1,0 +1,11 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.DotNet.ImageBuilder.Commands
+{
+    public class CopyDockerHubBaseImagesOptions : CopyImagesOptions
+    {
+        protected override string CommandHelp => "Copies external base images from Docker Hub to ACR";
+    }
+}
